% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_tab.r
\name{f_tab}
\alias{f_tab}
\title{Frequency distribution as a data.frame}
\usage{
f_tab(x, w = NULL)
}
\arguments{
\item{x}{Numeric, character or factor.}

\item{w}{Optional weight vector.}
}
\description{
Returns a \emph{data.frame} where each row shows the
frequencies of unique values in a vector. Weighting is supported.
}
\examples{
x <- sample(1:20, 100, replace = T)
f_tab(x)
}

