% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmedian.r
\name{gmedian}
\alias{gmedian}
\alias{gmedian.factor}
\alias{gmedian.numeric}
\title{Estimates Median of a grouped frequency distribution.}
\usage{
gmedian(x, percentile = 0.5, scale_interval = 1, w = NULL)

\method{gmedian}{factor}(x, percentile = 0.5, scale_interval = 1,
  w = NULL)

\method{gmedian}{numeric}(x, percentile = 0.5, scale_interval = 1,
  w = NULL)
}
\arguments{
\item{x}{factor or numeric vector to draw the frequency distribution from.}

\item{percentile}{value between 0 and 1; default = 0.5 (median)}

\item{scale_interval}{group intervall; default = 1}

\item{w}{an optional vector of weights}
}
\value{
Estimated Median for the given grouped frequency distribution as a
numeric vector.
}
\description{
\code{gmedian} calculates the interpolated median of a grouped
frequency distribution.
}
\details{
\code{gmedian} can be used with \code{factor} and \code{numeric}
vectors.
If the values of \code{x} represent numeric ranges, like 1-10, 11-20, 21-30,
values should be \code{numeric} starting values of range and
\code{scale_interval} should be used to indicate the range intervall.
}
\examples{
lvls <- c("very important", "2", "3", "4", "5", "6", "not important")
x <- sample(lvls, size = 100, replace = T)
gmedian(x)

# 25\% Quartile
gmedian(x, 0.25)
}

