% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmedian.r
\name{mQ}
\alias{intQR}
\alias{mQ}
\title{Interquartile range of a grouped frequency distribution.}
\usage{
mQ(x, ...)

intQR(x, ...)
}
\arguments{
\item{x}{factor or numeric vector to draw the frequency distribution from.}

\item{...}{additional parameters to pass to \code{gmedian()}}
}
\description{
Calculates interquartile range of a grouped frequency
distribution.
}
\section{Functions}{
\itemize{
\item \code{mQ}: Calculates halfed interquartile range
[ger. mittlerer Quartilsabstand] of a grouped frequency distribution
}}

