% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_compare_groups.r
\name{rs_compare_groups}
\alias{rs_compare_groups}
\title{Compare Medians of a questionnaire battery of equally coded ordinal/
categorial variables}
\usage{
rs_compare_groups(df, groups, item_labels = NULL, w = NULL, ...)
}
\arguments{
\item{df}{dataframe containing the rating scale variables}

\item{groups}{vector defining groups/ independent variable}

\item{item_labels}{optional character vector of variable labels}

\item{w}{Optional weight vector.}

\item{...}{additional options for gmedian(), e.g. percentile, scale_intervall or w(eight).}
}
\description{
Returns a \emph{data.frame} where each row represents one variable
and the columns show the median values of the comparison groups. An additional
colum shows the significance levels, based on the \emph{Kruskal-Wallis Rank
Sum Test}.
}
\examples{
# Create sample data.
df <- data.frame(replicate(6,sample(1:7, 100, replace = T)))

# Compare medians between groups accross several variables.
g <- sample(c("G1", "G2", "G3"), 100, replace = T)
res_comp <- rs_compare_groups(df = df, groups = g)
}

