% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_summary.r
\name{rs_summary}
\alias{rs_summary}
\title{Summarize a questionaire battery of equally coded ordinal/ categorial variables}
\usage{
rs_summary(df, item_labels = NULL, item_levels = NULL, l_mar = 12,
  cor = F, ...)
}
\arguments{
\item{df}{data.frame containing the variables}

\item{item_labels}{vector of variable names}

\item{item_levels}{vector of variable levels}

\item{l_mar}{numeric value for left margin of plot}
}
\description{
Returns a \emph{data.frame} where each row
represents one variable. It provides categories' frequencies,
N, median, and the interquartile range.
}
\examples{
# Create sample data.
df <- data.frame(replicate(6,sample(1:7, 100, replace = T)))

res <- rs_summary(df = df)
}

