% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vert_line_base.r, R/vert_line_gg.r
\name{vert_line_base}
\alias{vert_line_base}
\alias{vert_line_gg}
\title{Dotcharts of categorial variables}
\usage{
vert_line_base(x, max_val = NULL, min_val = 1, l_mar = 10, b_mar = 6,
  scale_interval = 1, ...)

vert_line_gg(x, max_val = NULL, min_val = 1)
}
\arguments{
\item{x}{data.frame containing the variables}

\item{max_val}{Absolute scale end as a numeric value.}

\item{min_val}{Absolute scale start as a numeric value; default = 1}

\item{l_mar}{Left margin; default = 10}

\item{b_mar}{Bottom margin; default = 6}

\item{scale_interval}{default = 1}
}
\description{
Dotcharts of categorial variables
}
\section{Functions}{
\itemize{
\item \code{vert_line_gg}: Uses ggplot2 for vertical line plot. Allows
for usage of ggplot2's theming.
}}
\examples{
# Create sample data.
df <- data.frame(replicate(6,sample(1:7, 100, replace = T)))

# Plot medians of several variables.
res <- rs_summary(df = df)
# Using base plot.
vert_line_base(x = res["Median"])
# Using ggplot2.
vert_line_gg(x = res["Median"])

# Compare medians between groups accross several variables.
g <- sample(c("G1", "G2", "G3"), 100, replace = T)
res_comp <- rs_compare_groups(df = df, groups = g)
# Using base plot.
vert_line_base(x = res_comp)
# Using ggplot2.
vert_line_gg(x = res_comp)
}

