% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_tab.r
\name{x_tab}
\alias{x_tab}
\title{Frequency distribution as a data.frame}
\usage{
x_tab(x, y, weight = NULL, p_max = 0.05)
}
\arguments{
\item{x}{Numeric, character or factor vector, defining the dependent variable.}

\item{y}{Numeric, character or factor vector, defining groups/ independent variable.}

\item{w}{Optional weight vector.}
}
\description{
Returns a \emph{data.frame} as a crosstable with indication of
statistic significance using chi-square. Weighting is supported.
}
\examples{
# Create sample data.
df <- data.frame(replicate(2,sample(1:7, 100, replace = T)))

x_tab(df[[1]], df[[2]])
}

